requirements.txt
google-genai>=1.0
python-dotenv>=1.0
fastapi>=0.115
uvicorn[standard]>=0.32
